Agp 156 search result (#33)
* transformed results js to tsx

* transformed results js to tsx

* added readonly to properties

* removed declared variables in the render method

* variable renamed
