+++ src/search/types.ts
@@ -1,0 +1,2 @@
+export type Company = {name: string, symbol:string};
+export type CompanySelect = (company:Company) => void;
